Consolidate UI types
