--- src/@types/fight-command.ts
@@ -1,3 +1,0 @@
-import type { BattleCommand } from "#enums/battle-command";
-
-export type FightCommand = typeof BattleCommand.FIGHT | typeof BattleCommand.TERA;
