--- src/@types/party-modifier-transfer-select-callback.ts
@@ -1,6 +1,0 @@
-export type PartyModifierTransferSelectCallback = (
-  fromCursor: number,
-  index: number,
-  itemQuantity?: number,
-  toCursor?: number,
-) => void;
