--- src/@types/party-select-callback.ts
@@ -1,3 +1,0 @@
-import type { PartyOption } from "#enums/party-option";
-
-export type PartySelectCallback = (cursor: number, option: PartyOption) => void;
