--- src/@types/pokemon-move-select-filter.ts
@@ -1,3 +1,0 @@
-import type { PokemonMove } from "#field/pokemon-move";
-
-export type PokemonMoveSelectFilter = (pokemonMove: PokemonMove) => string | null;
