--- src/@types/pokemon-select-filter.ts
@@ -1,3 +1,0 @@
-import type { PlayerPokemon } from "#field/player-pokemon";
-
-export type PokemonSelectFilter = (pokemon: PlayerPokemon) => string | null;
