--- src/@types/pokemon-species-filter.ts
@@ -1,3 +1,0 @@
-import type PokemonSpecies from "#data/pokemon-species";
-
-export type PokemonSpeciesFilter = (species: PokemonSpecies) => boolean;
