+++ src/battle-scene.ts
@@ -119,8 +119,8 @@ import type TrainerData from "#system/trainer-data";
 import { type Voucher, vouchers } from "#system/voucher";
 import { allTrainerConfigs } from "#trainer-configs/all-trainer-configs";
 import type { HeldModifierConfig, ModifierPredicate } from "#types/modifiers-types";
-import type { PokemonSpeciesFilter } from "#types/pokemon-species-filter";
 import type { AnySettingKey, SettingsUpdateEventArgs } from "#types/settings";
+import type { PokemonSpeciesFilter } from "#types/ui-types";
 import type { AbstractConstructor } from "#types/utility-types";
 import { AbilityBar } from "#ui/ability-bar";
 import { ArenaFlyout } from "#ui/arena-flyout";
