+++ src/data/mystery-encounters/utils/encounter-phase-utils.ts
@@ -53,7 +53,7 @@ import type MysteryEncounterOption from "#mystery-encounters/mystery-encounter-o
 import type PokemonData from "#system/pokemon-data";
 import { allTrainerConfigs } from "#trainer-configs/all-trainer-configs";
 import type { HeldModifierConfig } from "#types/modifiers-types";
-import type { PokemonSelectFilter } from "#types/pokemon-select-filter";
+import type { PokemonSelectFilter } from "#types/ui-types";
 import type { OptionSelectItem, OptionSelectModeConfig } from "#ui/option-select-config";
 import type { OptionSelectUiHandler } from "#ui/option-select-ui-handler";
 import type { PartyUiHandler } from "#ui/party-ui-handler";
