+++ src/data/trainer-config.ts
@@ -15,7 +15,7 @@ import { TrainerVariant } from "#enums/trainer-variant";
 import type { EnemyPokemon } from "#field/enemy-pokemon";
 import type { PersistentModifier } from "#modifier/modifier";
 import type { ModifierTypeFunc } from "#modifier/modifier-type";
-import type { PokemonSpeciesFilter } from "#types/pokemon-species-filter";
+import type { PokemonSpeciesFilter } from "#types/ui-types";
 import { clamp, coerceArray } from "#utils/common-utils";
 import { getPokemonSpecies } from "#utils/pokemon-utils";
 import { randItem, randSeedItem } from "#utils/random-utils";
