+++ src/modifier/modifier-type.ts
@@ -67,8 +67,7 @@ import { modifierPool } from "#modifier/modifier-pools";
 import { modifierTypes } from "#modifier/modifier-types";
 import { settings } from "#system/settings-manager";
 import { getVoucherTypeIcon, getVoucherTypeName } from "#system/voucher";
-import type { PokemonMoveSelectFilter } from "#types/pokemon-move-select-filter";
-import type { PokemonSelectFilter } from "#types/pokemon-select-filter";
+import type { PokemonMoveSelectFilter, PokemonSelectFilter } from "#types/ui-types";
 import type { EnumValues } from "#types/utility-types";
 import { getModifierTierTextTint } from "#ui/text-utils";
 import { getBerryEffectDescription, getBerryName } from "#utils/berry-utils";
