+++ src/phases/command-phase.ts
@@ -22,8 +22,8 @@ import { UiMode } from "#enums/ui-mode";
 import type { Pokemon } from "#field/pokemon";
 import { getMoveTargets, type MoveTargetSet } from "#moves/move";
 import { FieldPhase } from "#phases/base/field-phase";
-import type { FightCommand } from "#types/fight-command";
 import type { TurnMove } from "#types/move-types";
+import type { FightCommand } from "#types/ui-types";
 import type { CommandUiHandler } from "#ui/command-ui-handler";
 import type { FightUiHandler } from "#ui/fight-ui-handler";
 import { isNil } from "#utils/common-utils";
